usr.sbin/bhyve: Introduce block_if mux, and block_local
