+++ usr.sbin/bhyve/Makefile
@@ -1,5 +1,5 @@
 #
-# $FreeBSD$
+# $FreeBSD: head/usr.sbin/bhyve/Makefile 349803 2019-07-07 12:15:24Z vmaffione $
 #
 
 .include <src.opts.mk>
@@ -20,6 +20,7 @@ SRCS=	\
 	bhyvegc.c		\
 	bhyverun.c		\
 	block_if.c		\
+	block_local.c		\
 	bootrom.c		\
 	console.c		\
 	consport.c		\
